pyproject.toml
[project]
name = "geolocation"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Jett Chen", email = "jettchen12345@gmail.com" }
]
dependencies = [
    "numpy>=1.26.4",
    "folium>=0.16.0",
    "geopy>=2.4.1",
    "langchain>=0.1.11",
    "markdownify>=0.11.6",
    "openai>=1.13.3",
    "overpy>=0.7",
    "pillow>=10.2.0",
    "pydantic>=2.6.3",
    "python-dotenv>=1.0.1",
    "requests>=2.31.0",
    "rich>=13.7.1",
    "streetview>=0.0.6",
    "tqdm>=4.66.2",
    "aiohttp>=3.9.3",
    "albumentations>=1.4.1",
    "backoff>=2.2.1",
    "cffi>=1.16.0",
    "functiontrace>=0.3.6",
    "gdown>=5.1.0",
    "mpmath>=1.3.0",
    "opencv-python>=4.9.0.80",
    "overrides>=7.7.0",
    "pandas>=2.2.1",
    "pytz>=2024.1",
    "scikit-learn>=1.4.1.post1",
    "scipy>=1.12.0",
    "setuptools>=69.1.1",
    "sympy>=1.12",
    "timm>=0.9.16",
    "transformers>=4.38.2",
    "typing-extensions>=4.10.0",
    "attrs>=23.2.0",
    "colorama>=0.4.6",
    "fsspec>=2024.2.0",
    "markupsafe>=2.1.5",
    "networkx>=3.2.1",
    "pynvml>=11.5.0",
    "python-dateutil>=2.9.0.post0",
    "pyyaml>=6.0.1",
    "safetensors>=0.4.2",
    "six>=1.16.0",
    "sqlalchemy>=2.0.28",
    "yarl>=1.9.4",
    "torch>=2.2.1",
    "torchvision>=0.17.1",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/geolocation"]

[project.scripts]
agent = "agent:main"
